utils/fastq_utils: migrate 'AnalysisFastq' to new 'IlluminaFastqAttrs'.
Moves the functionality of the utils.AnalysisFastq class to a new class
fastq_utils.IlluminaFastqAttrs (also moves the BaseFastqAttrs class to the
fastq_utils module).

utils.AnalysisFastq is now a wrapper for fastq_utils.IlluminaFastqAttrs.
