+++ auto_process_ngs/test/test_utils.py
@@ -11,282 +11,8 @@ from bcftbx.utils import find_program
 from auto_process_ngs.mock import MockAnalysisDirFactory
 from auto_process_ngs.mock import MockAnalysisProject
 from auto_process_ngs.applications import Command
+from auto_process_ngs.fastq_utils import BaseFastqAttrs
 from auto_process_ngs.utils import *
-
-class TestAnalysisFastq(unittest.TestCase):
-    """Tests for the AnalysisFastq class
-
-    """
-
-    def test_full_name(self):
-        """Handle full Illumina-style fastq name
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_ACAGTG_L003_R2_001')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,
-                         'NH1_ChIP-seq_Gli1_ACAGTG_L003_R2_001')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,'ACAGTG')
-        self.assertEqual(fq.lane_number,3)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,1)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_ACAGTG_L003_R2_001')
-
-    def test_full_name_dual_index(self):
-        """Handle full Illumina-style fastq name with dual index
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_ACAGTG-GTTCAC_L003_R2_001')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,
-                         'NH1_ChIP-seq_Gli1_ACAGTG-GTTCAC_L003_R2_001')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,'ACAGTG-GTTCAC')
-        self.assertEqual(fq.lane_number,3)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,1)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_ACAGTG-GTTCAC_L003_R2_001')
-
-    def test_full_name_blc2fastq2(self):
-        """Handle Illumina fastq name from bcl2fastq2
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_S4_L003_R2_001')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,
-                         'NH1_ChIP-seq_Gli1_S4_L003_R2_001')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,4)
-        self.assertEqual(fq.barcode_sequence,None)
-        self.assertEqual(fq.lane_number,3)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,1)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_S4_L003_R2_001')
-
-    def test_index_read_blc2fastq2(self):
-        """Handle Illumina index read fastq name from bcl2fastq2
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_S4_L003_I1_001')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.sample_number,4)
-        self.assertEqual(fq.barcode_sequence,None)
-        self.assertEqual(fq.lane_number,3)
-        self.assertEqual(fq.read_number,1)
-        self.assertEqual(fq.set_number,1)
-        self.assertTrue(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_S4_L003_I1_001')
-
-    def test_name_no_lane_blc2fastq2(self):
-        """Handle Illumina fastq name from bcl2fastq2 (without lane)
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_S4_R2_001')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,
-                         'NH1_ChIP-seq_Gli1_S4_R2_001')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,4)
-        self.assertEqual(fq.barcode_sequence,None)
-        self.assertEqual(fq.lane_number,None)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,1)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_S4_R2_001')
-
-    def test_name_only(self):
-        """Handle reduced fastq name (sample name only)
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,None)
-        self.assertEqual(fq.lane_number,None)
-        self.assertEqual(fq.read_number,None)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1')
-
-    def test_name_only_paired_end(self):
-        """Handle reduced fastq name (sample name only, paired end)
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_R2')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq_Gli1_R2')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,None)
-        self.assertEqual(fq.lane_number,None)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_R2')
-
-    def test_name_and_lane(self):
-        """Handle reduced fastq name (sample name and lane)
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_L001')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq_Gli1_L001')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,None)
-        self.assertEqual(fq.lane_number,1)
-        self.assertEqual(fq.read_number,None)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_L001')
-
-    def test_name_and_lane_paired_end(self):
-        """Handle reduced fastq name (sample name and lane, paired end)
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_L001_R2')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq_Gli1_L001_R2')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,None)
-        self.assertEqual(fq.lane_number,1)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_L001_R2')
-
-    def test_name_and_tag(self):
-        """Handle reduced fastq name (sample name and barcode)
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_ACAGTG')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq_Gli1_ACAGTG')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,'ACAGTG')
-        self.assertEqual(fq.lane_number,None)
-        self.assertEqual(fq.read_number,None)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_ACAGTG')
-
-    def test_name_and_tag_paired_end(self):
-        """Handle reduced fastq name (sample name and barcode, paired end)
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_ACAGTG_R2')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq_Gli1_ACAGTG_R2')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,'ACAGTG')
-        self.assertEqual(fq.lane_number,None)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_ACAGTG_R2')
-
-    def test_name_tag_and_lane(self):
-        """Handle reduced fastq name (sample name, barcode and lane)
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_ACAGTG_L001')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq_Gli1_ACAGTG_L001')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,'ACAGTG')
-        self.assertEqual(fq.lane_number,1)
-        self.assertEqual(fq.read_number,None)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_ACAGTG_L001')
-
-    def test_name_tag_and_lane_paired_end(self):
-        """Handle reduced fastq name (sample name, barcode and lane, paired end)
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq_Gli1_ACAGTG_L001_R2')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq_Gli1_ACAGTG_L001_R2')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,'ACAGTG')
-        self.assertEqual(fq.lane_number,1)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_ACAGTG_L001_R2')
-
-    def test_AGTC_sample_names(self):
-        """Handle sample names consisting of letters 'A', 'G', 'T' and 'C'
-        """
-        for name in ('A','G','T','C','AGCT'):
-            fq = AnalysisFastq('%s_R1' % name)
-            self.assertEqual(fq.sample_name,name)
-            self.assertEqual(fq.sample_number,None)
-            self.assertEqual(fq.barcode_sequence,None)
-            self.assertEqual(fq.lane_number,None)
-            self.assertEqual(fq.read_number,1)
-            self.assertEqual(fq.set_number,None)
-            self.assertFalse(fq.is_index_read)
-            self.assertEqual(str(fq),'%s_R1' % name)
-
-    def test_non_standard_sample_name(self):
-        """Handle non-standard Fastq names with sample name only
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq.r2')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq.r2')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,None)
-        self.assertEqual(fq.lane_number,None)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq.r2')
-
-    def test_non_standard_sample_name_with_dots(self):
-        """Handle non-standard Fastq names with sample name containing dots
-        """
-        fq = AnalysisFastq('NH1.2.r2')
-        self.assertEqual(fq.sample_name,'NH1.2')
-        self.assertEqual(fq.basename,'NH1.2.r2')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,None)
-        self.assertEqual(fq.lane_number,None)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,None)
-        self.assertEqual(str(fq),'NH1.2.r2')
-
-    def test_non_standard_sample_name_and_barcode(self):
-        """Handle non-standard Fastq names with sample name and barcode
-        """
-        fq = AnalysisFastq('NH1_ChIP-seq.ACAGTG.r2')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq.ACAGTG.r2')
-        self.assertEqual(fq.extension,'')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,'ACAGTG')
-        self.assertEqual(fq.lane_number,None)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,None)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq.ACAGTG.r2')
-
-    def test_input_is_full_path(self):
-        """Handle input as full path to Fastq file
-        """
-        fq = AnalysisFastq('/data/Project_NH/Sample_NH1/NH1_ChIP-seq_Gli1_ACAGTG_L003_R2_001.fastq.gz')
-        self.assertEqual(fq.sample_name,'NH1_ChIP-seq_Gli1')
-        self.assertEqual(fq.basename,'NH1_ChIP-seq_Gli1_ACAGTG_L003_R2_001')
-        self.assertEqual(fq.extension,'.fastq.gz')
-        self.assertEqual(fq.sample_number,None)
-        self.assertEqual(fq.barcode_sequence,'ACAGTG')
-        self.assertEqual(fq.lane_number,3)
-        self.assertEqual(fq.read_number,2)
-        self.assertEqual(fq.set_number,1)
-        self.assertFalse(fq.is_index_read)
-        self.assertEqual(str(fq),'NH1_ChIP-seq_Gli1_ACAGTG_L003_R2_001')
 
 class TestAnalysisDir(unittest.TestCase):
     """Tests for the AnalysisDir class
