+++ auto_process_ngs/utils.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 #
 #     utils: utility classes & funcs for auto_process_ngs module
-#     Copyright (C) University of Manchester 2013-2017 Peter Briggs
+#     Copyright (C) University of Manchester 2013-2018 Peter Briggs
 #
 ########################################################################
 #
@@ -20,7 +20,6 @@ tree at some point.
 
 Classes:
 
-- BaseFastqAttrs
 - AnalysisFastq:
 - AnalysisDir:
 - AnalysisProject:
@@ -60,13 +59,13 @@ import bcftbx.IlluminaData as IlluminaData
 import bcftbx.JobRunner as JobRunner
 import bcftbx.Pipeline as Pipeline
 import bcftbx.utils as bcf_utils
-from bcftbx.qc.report import strip_ngs_extensions
 from bcftbx.Md5sum import md5sum
 from .metadata import AnalysisDirMetadata
 from .metadata import AnalysisDirParameters
 from .metadata import AnalysisProjectInfo
 from .metadata import ProjectMetadataFile
 from .metadata import AnalysisProjectQCDirInfo
+from .fastq_utils import IlluminaFastqAttrs
 from qc.illumina_qc import QCReporter
 from qc.illumina_qc import QCSample
 from qc.illumina_qc import expected_qc_outputs
@@ -79,262 +78,10 @@ logger = logging.getLogger(__name__)
 # Classes
 #######################################################################
 
-class BaseFastqAttrs(object):
+class AnalysisFastq(IlluminaFastqAttrs):
     """
-    Base class for extracting information about a Fastq file
-
-    Instances of this class provide the follow attributes:
-
-    fastq:            the original fastq file name
-    basename:         basename with NGS extensions stripped
-    extension:        full extension e.g. '.fastq.gz'
-    sample_name:      name of the sample
-    sample_number:    integer (or None if no sample number)
-    barcode_sequence: barcode sequence (string or None)
-    lane_number:      integer (or None if no lane number)
-    read_number:      integer (or None if no read number)
-    set_number:       integer (or None if no set number)
-    is_index_read:    boolean (True if index read, False if not)
-
-    Subclasses should process the supplied Fastq name and set
-    these attributes appropriately.
+    Wrapper for IlluminaFastqAttrs
     """
-    def __init__(self,fastq):
-        # Store name
-        self.fastq = fastq
-        # Basename and extension
-        self.basename = strip_ngs_extensions(self.fastq)
-        self.extension = self.fastq[len(self.basename):]
-        self.basename = os.path.basename(self.basename)
-        # Values that should be derived from the name
-        # (should be set by subclass)
-        self.sample_name = None
-        self.sample_number = None
-        self.barcode_sequence = None
-        self.lane_number = None
-        self.read_number = None
-        self.set_number = None
-        self.is_index_read = False
-    def __repr__(self):
-        return self.basename
-
-class AnalysisFastq(BaseFastqAttrs):
-    """Class for extracting information about Fastq files
-
-    Given the name of a Fastq file, extract data about the sample name,
-    barcode sequence, lane number, read number and set number.
-
-    The name format can be a 'full' Fastq name as generated by CASAVA or
-    bcl2fastq 1.8, which follows the general form:
-
-    <sample_name>_<barcode_sequence>_L<lane_number>_R<read_number>_<set_number>.fastq.gz
-
-    e.g. for
-
-    NA10831_ATCACG_L002_R1_001.fastq.gz
-
-    sample_name = 'NA10831_ATCACG_L002_R1_001'
-    barcode_sequence = 'ATCACG'
-    lane_number = 2
-    read_number = 1
-    set_number = 1
-
-    Alternatively it can be a full Fastq name as generated by bcl2fastq2,
-    of the general form:
-
-    <sample_name>_S<sample_number>_L<lane_number>_R<read_number>_001.fastq.gz
-
-    e.g. for
-
-    ES_exp1_S4_L003_R2_001.fastq.gz
-
-    sample_name = 'ES_exp1'
-    sample_number = 4
-    lane_number = 3
-    read_number = 2
-    set_number = 1
-
-    bcl2fastq can also produce 'index read' Fastq files where the
-    R1/R2 is replaced by I1, e.g.:
-
-    ES_exp1_S4_L003_I1_001.fastq.gz
-
-    Alternatively it can be a 'reduced' version where one or more
-    of the components has been omitted (typically because they are
-    redundant in uniquely distinguishing a Fastq file within a
-    set of Fastqs).
-
-    The reduced formats are:
-
-    <sample_name>
-    <sample_name>_L<lane_number>
-    <sample_name>_<barcode_sequence>
-    <sample_name>_<barcode_sequence>_L<lane_number>
-
-    with an optional suffix '_R<read_number>' for paired end sets.
-
-    e.g.
-
-    NA10831
-    NA10831_L002
-    NA10831_ATCACG
-    NA10831_ATCACG_L002
-
-    Finally, the name can be a non-standard name of the form:
-
-    <sample_name>.r<read_number>
-
-    or
-
-    <sample_name>.<barcode_sequence>.r<read_number>
-
-    In this case the sample_names are permitted to include dots.
-
-    Provides the follow attributes:
-
-    fastq:            the original fastq file name
-    sample_name:      name of the sample (leading part of the name)
-    sample_number:    integer (or None if no sample number)
-    barcode_sequence: barcode sequence (string or None)
-    lane_number:      integer (or None if no lane number)
-    read_number:      integer (or None if no read number)
-    set_number:       integer (or None if no set number)
-    is_index_read:    boolean (True if index read, False if not)
-
-    """
-
-    def __init__(self,fastq):
-        """Create and populate a new AnalysisFastq object
-
-        Arguments:
-          fastq: name of the fastq.gz (optionally can include leading path)
-
-        """
-        BaseFastqAttrs.__init__(self,fastq)
-        # Base name for sample (no leading path or extension)
-        fastq_base = self.basename
-        # Determine if it's a non-standard (dot-separated) name
-        #
-        # These are names of the form e.g.
-        # NH1.2.r2
-        # or
-        # NH1_ChIP-seq.ACAGTG.r2
-        if fastq_base.count('.') > 0:
-            self.delimiter = '.'
-            fields = fastq_base.split('.')
-            field = fields[-1]
-            if len(field) == 2 and field.startswith('r'):
-                # Read number
-                self.read_number = int(field[1])
-                fields = fields[:-1]
-                field = fields[-1]
-            if len(fields) > 1:
-                # Barcode sequence
-                is_tag = True
-                for f in field.split('-'):
-                    for c in f:
-                        is_tag = is_tag and c in 'ACGTN'
-                if is_tag:
-                    self.barcode_sequence = field
-                    fields = fields[:-1]
-                    field = fields[-1]
-            # Remaining fields are the sample name
-            self.sample_name = '.'.join(fields)
-            assert(self.sample_name != '')
-            return
-        # Some form of Illumina-derived name
-        #
-        # Full Illumina-style names are e.g.
-        # NH1_ChIP-seq_Gli1_ACAGTG_L001_R1_001
-        # or
-        # NH1_ChIP-seq_Gli1_S4_L003_R2_001
-        #
-        # We have shorter name formats where redundant parts are
-        # omitted, the patterns are:
-        # NAME          e.g. NH1_ChIP-seq_Gli1
-        # NAME+LANE     e.g. NH1_ChIP-seq_Gli1_L001
-        # NAME+TAG      e.g. NH1_ChIP-seq_Gli1_ACAGTG
-        # NAME+TAG+LANE e.g. NH1_ChIP-seq_Gli1_ACAGTG_L001
-        #
-        # Also read number (i.e. R1 or R2) is appended but only for
-        # paired end samples
-        #
-        # The set number is never included, except for full names
-        fields = fastq_base.split('_')
-        self.delimiter = '_'
-        # Deal with set number first e.g. 001
-        field = fields[-1]
-        ##logger.debug("Test for set number %s" % field)
-        if len(field) == 3 and field.isdigit():
-            self.set_number = int(field)
-            fields = fields[:-1]
-        # Deal with trailing read number e.g. R1
-        field = fields[-1]
-        ##logger.debug("Test for read number %s" % field)
-        if len(field) == 2:
-            if field.startswith('R'):
-                self.read_number = int(field[1])
-                fields = fields[:-1]
-            elif field.startswith('I'):
-                self.read_number = int(field[1])
-                self.is_index_read = True
-                fields = fields[:-1]
-        # Deal with trailing lane number e.g. L001
-        field = fields[-1]
-        ##logger.debug("Test for lane number %s" % field)
-        if len(field) == 4 and field.startswith('L') and field[1:].isdigit():
-            self.lane_number = int(field[1:])
-            fields = fields[:-1]
-        # Deal with trailing index tag e.g. ATTGCT or ATTGCT-CCTAAG
-        field = fields[-1]
-        ##logger.debug("Test for barcode sequence %s" % field)
-        if len(fields) > 1:
-            # This mustn't be the last field: if it is then it's
-            # not the tag - it's the name
-            is_tag = True
-            for f in field.split('-'):
-                for c in f:
-                    is_tag = is_tag and c in 'ACGTN'
-            if is_tag:
-                self.barcode_sequence = field
-                fields = fields[:-1]
-                ##logger.debug("Identified barcode sequence as %s" % self.barcode_sequence)
-            else:
-                # Alternatively might be the sample number
-                if field.startswith('S'):
-                    try:
-                        if field[1:].isdigit():
-                            self.sample_number = int(field[1:])
-                            fields = fields[:-1]
-                    except IndexError:
-                        pass
-        # What's left is the name
-        ##logger.debug("Remaining fields: %s" % fields)
-        self.sample_name = '_'.join(fields)
-        assert(self.sample_name != '')
-
-    def __repr__(self):
-        """Implement __repr__ built-in
-
-        """
-        # Reconstruct name
-        fq = ["%s" % self.sample_name]
-        if self.sample_number is not None:
-            fq.append("S%d" % self.sample_number)
-        if self.barcode_sequence is not None:
-            fq.append("%s" % self.barcode_sequence)
-        if self.lane_number is not None:
-            fq.append("L%03d" % self.lane_number)
-        if self.read_number is not None:
-            if self.delimiter == '.':
-                fq.append("r%d" % self.read_number)
-            elif self.is_index_read:
-                fq.append("I%d" % self.read_number)
-            else:
-                fq.append("R%d" % self.read_number)
-        if self.set_number is not None:
-            fq.append("%03d" % self.set_number)
-        return self.delimiter.join(fq)
 
 class AnalysisDir:
     """Class describing an analysis directory
